// src/app/shared/models/introduction.ts
export interface iIntroduction {
  firstName: string;
  lastName: string;
  fullName: string;
  titles: string[];
  photo_Intro: string;
    // title: string;
    // description: string;
    // profileImageUrl: string;
    // socialLinks: {
    //     github?: string;
    //     linkedin?: string;
    //     twitter?: string;
    //     [key: string]: string | undefined;
    // };
}